--- a/hw10_program_optimization/model/user.go
+++ b/hw10_program_optimization/model/user.go
@@ -1,0 +1,5 @@
+package model
+
+type User struct {
+	Email string `json:"Email,nocopy"` //nolint:staticcheck,tagliatelle
+}