--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -1,9 +1,11 @@
+//go:build !bench
 // +build !bench
 
 package hw10programoptimization
 
 import (
 	"bytes"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -36,4 +38,26 @@ func TestGetDomainStat(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
 	})
+}
+
+func TestEmptyDomain(t *testing.T) {
+	_, err := GetDomainStat(strings.NewReader(""), "")
+
+	require.Error(t, err)
+}
+
+func TestEmptyReader(t *testing.T) {
+	res, err := GetDomainStat(strings.NewReader(""), "gov")
+
+	require.Empty(t, res)
+	require.NoError(t, err)
+}
+
+func TestInvalidEmail(t *testing.T) {
+	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@bro@Wsecat.gov","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}`
+
+	_, err := GetDomainStat(bytes.NewBufferString(data), "gov")
+
+	require.ErrorContains(t, err, "invalid email")
 }